자신의 정보만 삭제 가능

+++ backend/src/main/java/com/example/backend/controller/member/MemberController.java
@@ -57,17 +57,25 @@ public class MemberController {
     }
 
     @DeleteMapping("remove")
-    public ResponseEntity<Map<String, Object>> remove(@RequestBody Member member) {
-        if (service.remove(member)) {
-            //잘됨
-            return ResponseEntity.ok(Map.of("message",
-                    Map.of("type", "success",
-                            "text", "회원정보를 삭제하였습니다.")));
+    @PreAuthorize("isAuthenticated()")
+    public ResponseEntity<Map<String, Object>> remove(
+            @RequestBody Member member,
+            Authentication authentication) {
+        if (service.hasAccess(member.getId(), authentication)) {
+
+            if (service.remove(member)) {
+                //잘됨
+                return ResponseEntity.ok(Map.of("message",
+                        Map.of("type", "success",
+                                "text", "회원정보를 삭제하였습니다.")));
+            } else {
+                return ResponseEntity.badRequest()
+                        .body(Map.of("message",
+                                Map.of("type", "warning",
+                                        "text", "정확한 정보를 입력해주세요.")));
+            }
         } else {
-            return ResponseEntity.badRequest()
-                    .body(Map.of("message",
-                            Map.of("type", "warning",
-                                    "text", "정확한 정보를 입력해주세요.")));
+            return ResponseEntity.status(403).build();
         }
     }
